+++ apps/api/src/app/export/export.module.ts
@@ -2,6 +2,7 @@ import { AccountModule } from '@ghostfolio/api/app/account/account.module';
 import { OrderModule } from '@ghostfolio/api/app/order/order.module';
 import { TransformDataSourceInRequestModule } from '@ghostfolio/api/interceptors/transform-data-source-in-request/transform-data-source-in-request.module';
 import { ApiModule } from '@ghostfolio/api/services/api/api.module';
+import { MarketDataModule } from '@ghostfolio/api/services/market-data/market-data.module';
 import { TagModule } from '@ghostfolio/api/services/tag/tag.module';
 
 import { Module } from '@nestjs/common';
@@ -14,6 +15,7 @@ import { ExportService } from './export.service';
   imports: [
     AccountModule,
     ApiModule,
+    MarketDataModule,
     OrderModule,
     TagModule,
     TransformDataSourceInRequestModule