+++ libs/common/src/lib/interfaces/export.interface.ts
@@ -1,4 +1,11 @@
-import { Account, Order, Platform, Tag } from '@prisma/client';
+import {
+  Account,
+  DataSource,
+  Order,
+  Platform,
+  SymbolProfile,
+  Tag
+} from '@prisma/client';
 
 import { AccountBalance } from './account-balance.interface';
 
@@ -15,7 +22,10 @@ export interface Export {
     | 'symbolProfileId'
     | 'updatedAt'
     | 'userId'
-  > & { date: string; symbol: string })[];
+  > & { dataSource: DataSource; date: string; symbol: string })[];
+  assetProfiles: (Omit<SymbolProfile, 'createdAt' | 'updatedAt' | 'userId'> & {
+    marketData: { date: string; marketPrice: number }[];
+  })[];
   meta: {
     date: string;
     version: string;